Work on sign-up page functionality
Completed events.js as it pertains to sign-up. Functions as expected.

Completed ui.js as it pertains to sign-up. Functions as expected.

Completed api.js as it pertains to sign-up. Functions as expected.

Added a success/error message to index.html & assigned ID 'messages' to
show a success or error for the submission. Will increase functionality
to other .sh functions such as sign-in, change-password, etc for
testing.

Added some class identifiers to index.scss to change font color of a
success message to green, and an error message to red.