Work sign-in module feature
Created functioning app.js code as it pertains to sign-in feature.

Created functioning events.js code as it pertains to sign-in feature.

Created functioning api.js code as it pertains to sign-in feature.

Created functioning ui.js code as it pertains to sign-in feature, which
will hide the sign up and sign in buttons + input fields and then show
some other data which hasn't yet been written, but will be specific to
someone who is already logged in. (essentially, remainder of the game).