Work index.scss re: auth/unauth id's
Added code to hide any of the code under the ID 'authenticated', so
features designed only for a logged-in user will not be available to
a logged out user. So far the only feature of this nature is to change
password, which has not yet been completed, only placeholder html for
the time being.