Work game API + UI functions
Added feature to index games (show all games for user) and the necessary
means to accomplish this.

Shell script tested in terminal functioning as expected.

Linked to game API.

UI button available to click, appears is functioning as expected and
presently returning a value of an empty games array as no games have
been created/saved to date for this user. Will be adding the ability
to create a game in the future.